Data-Ingestion-Extraction/pyproject.toml
[tool.poetry]
name = "csam_inventory"
version = "0.1.0"
description = ""
authors = ["Sudhangi Suthrave <sudhangi.suthrave@ed.gov>", "Arthur Neuman <arthur.neuman@ed.gov"]

[tool.poetry.dependencies]
python = "^3.8"
splinter =  "^0.15.0"
PyYAML = "^5.4.1"
beautifulsoup4 = "^4.9.3"
pandas = "^1.3.1"
openpyxl = "^3.0.7"
html5lib = "^1.1"
"pdfminer.six" = "^20201018"
pywin32 = "^301"

[tool.poetry.dev-dependencies]
ipython = "^7.26.0"
pylint = "^2.9.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
